Move Converter and Validator to own class.
